restructure directories going forward
