+++ s2wav/filter_factory/__init__.py
@@ -1,0 +1,3 @@
+from . import tiling
+from . import filters
+from . import kernels
